[Compiler Toolkit] Avoid DTensorize BlockMask for FlexAttention
