+++ torchtitan/experiments/compiler_toolkit/common_utils.py
@@ -6,6 +6,10 @@
 
 from contextlib import contextmanager
 
+import torch
+from torch.distributed.tensor import DTensor, Replicate
+from torch.utils._pytree import tree_map
+
 from torchtitan.config import JobConfig
 
 
@@ -18,3 +22,21 @@ def disable_compile(job_config: JobConfig):
         yield
     finally:
         job_config.compile.enable = original_value
+
+
+def parallelize_inputs(world_mesh, args, kwargs):
+    def to_dtensor(tensor):
+        if isinstance(tensor, torch.Tensor):
+            return DTensor.from_local(tensor, world_mesh["tp"], [Replicate()])
+        return tensor
+
+    dt_args = tree_map(to_dtensor, args)
+
+    # TODO: When using flex_attention, BlockMask would show up in kwargs,
+    # and it's unclear how to convert it to DTensor. If I use to_dtensor,
+    # it would fail with Dynamo Error: P2011360347
+    # dt_kwargs = tree_map(to_dtensor, kwargs)
+
+    dt_kwargs = kwargs
+
+    return dt_args, dt_kwargs
